fic: Refactored some of the changes that were lost by merging the latest S7 updates (Also gave the code a thorough review and refactored many details)
chore: Updated the "http" -> "https" in the apache header, where this was wrong.

+++ plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/optimizer/S7Optimizer.java
@@ -19,23 +19,25 @@
 package org.apache.plc4x.java.s7.readwrite.optimizer;
 
 import org.apache.plc4x.java.api.exceptions.PlcRuntimeException;
-import org.apache.plc4x.java.api.messages.*;
+import org.apache.plc4x.java.api.messages.PlcReadRequest;
+import org.apache.plc4x.java.api.messages.PlcWriteRequest;
 import org.apache.plc4x.java.api.model.PlcTag;
 import org.apache.plc4x.java.api.value.PlcValue;
 import org.apache.plc4x.java.s7.readwrite.*;
 import org.apache.plc4x.java.s7.readwrite.context.S7DriverContext;
+import org.apache.plc4x.java.s7.readwrite.tag.S7ClkTag;
+import org.apache.plc4x.java.s7.readwrite.tag.S7SzlTag;
 import org.apache.plc4x.java.s7.readwrite.tag.S7Tag;
-import org.apache.plc4x.java.s7.readwrite.MemoryArea;
-import org.apache.plc4x.java.s7.readwrite.TransportSize;
 import org.apache.plc4x.java.spi.context.DriverContext;
 import org.apache.plc4x.java.spi.messages.DefaultPlcReadRequest;
 import org.apache.plc4x.java.spi.messages.DefaultPlcWriteRequest;
 import org.apache.plc4x.java.spi.messages.utils.TagValueItem;
 import org.apache.plc4x.java.spi.optimizer.BaseOptimizer;
 
-import java.util.*;
-import org.apache.plc4x.java.s7.readwrite.tag.S7ClkTag;
-import org.apache.plc4x.java.s7.readwrite.tag.S7SzlTag;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.List;
 
 public class S7Optimizer extends BaseOptimizer {
 
@@ -64,13 +66,13 @@ public class S7Optimizer extends BaseOptimizer {
         LinkedHashMap<String, PlcTag> curTags = new LinkedHashMap<>();
 
         for (String tagName : readRequest.getTagNames()) {
-                    
-            if ((readRequest.getTag(tagName) instanceof S7SzlTag) || 
+
+            if ((readRequest.getTag(tagName) instanceof S7SzlTag) ||
                 (readRequest.getTag(tagName) instanceof S7ClkTag)) {
                 curTags.put(tagName, readRequest.getTag(tagName));
                 continue;
             }
-            
+
             S7Tag tag = (S7Tag) readRequest.getTag(tagName);
 
             int readRequestItemSize = S7_ADDRESS_ANY_SIZE;
@@ -121,13 +123,13 @@ public class S7Optimizer extends BaseOptimizer {
 
     @Override
     protected List<PlcWriteRequest> processWriteRequest(PlcWriteRequest writeRequest, DriverContext driverContext) {
-        
+
         for (String tagName : writeRequest.getTagNames()) {
             if (writeRequest.getTag(tagName) instanceof S7ClkTag) {
-                return Collections.singletonList( writeRequest);
+                return Collections.singletonList(writeRequest);
             }
-        } 
-        
+        }
+
         S7DriverContext s7DriverContext = (S7DriverContext) driverContext;
         List<PlcWriteRequest> processedRequests = new LinkedList<>();
 
@@ -138,14 +140,14 @@ public class S7Optimizer extends BaseOptimizer {
 
         // List of all items in the current request.
         LinkedHashMap<String, TagValueItem> curTags = new LinkedHashMap<>();
-        
+
         for (String tagName : writeRequest.getTagNames()) {
             S7Tag tag = (S7Tag) writeRequest.getTag(tagName);
             PlcValue value = writeRequest.getPlcValue(tagName);
 
             int writeRequestItemSize = S7_ADDRESS_ANY_SIZE + 4/* Size of Payload item header*/;
             if (tag.getDataType() == TransportSize.BOOL) {
-                writeRequestItemSize += Math.ceil((double) tag.getNumberOfElements() / 8);
+                writeRequestItemSize += (int) Math.ceil((double) tag.getNumberOfElements() / 8);
             } else {
                 writeRequestItemSize += (tag.getNumberOfElements() * tag.getDataType().getSizeInBytes());
             }
